fix: stub window.config before setupTests imports the app
`utils/countries` reads `window.config` while it is being evaluated, and
`setupTests.ts` pulls in the application through `tests/util`, so its own
`vi.stubGlobal('config', ...)` statement ran only after that import had
already thrown. Every client test failed to collect as a result.

Moving the stub into `setupConfig.ts` and listing it ahead of
`setupTests.ts` in `setupFiles` makes it run before any application module
is imported. `vi.hoisted` would have been the smaller change, but it does
not exist in the vitest 0.25.5 this branch pins.
