+++ packages/client/vite.config.ts
@@ -134,7 +134,7 @@ export default defineConfig(({ mode }) => {
     ],
     test: {
       environment: 'jsdom',
-      setupFiles: './src/setupTests.ts',
+      setupFiles: ['./src/setupConfig.ts', './src/setupTests.ts'],
       testTimeout: 60000,
       hookTimeout: 60000,
       globals: true
